// src/generated/prisma/models/FileAsset.ts
export type FileAssetWhereInput = {
  AND?: Prisma.FileAssetWhereInput | Prisma.FileAssetWhereInput[]
  OR?: Prisma.FileAssetWhereInput[]
  NOT?: Prisma.FileAssetWhereInput | Prisma.FileAssetWhereInput[]
  id?: Prisma.StringFilter<"FileAsset"> | string
  name?: Prisma.StringFilter<"FileAsset"> | string
  blobUrl?: Prisma.StringFilter<"FileAsset"> | string
  blobPathname?: Prisma.StringFilter<"FileAsset"> | string
  contentType?: Prisma.StringFilter<"FileAsset"> | string
  size?: Prisma.IntFilter<"FileAsset"> | number
  categoryId?: Prisma.StringFilter<"FileAsset"> | string
  createdAt?: Prisma.DateTimeFilter<"FileAsset"> | Date | string
  category?: Prisma.XOR<Prisma.FileCategoryScalarRelationFilter, Prisma.FileCategoryWhereInput>
}